room_slot/requirements.txt
Django==3.0.3
Pillow>=8.0.0
Djangorestframework==3.11.0
